feat: support theme detect on fb
